Module reorganisation. Triggers major version bump.
