+++ rhine/src/FRP/Rhine/Clock/Realtime/Audio.hs
@@ -26,10 +26,10 @@ import Control.Monad.IO.Class
 
 
 -- dunai
-import Control.Monad.Trans.MSF.Except
+import Control.Monad.Trans.MSF.Except hiding (step)
 
 -- rhine
-import FRP.Rhine
+import FRP.Rhine.Clock
 
 -- | Rates at which audio signals are typically sampled.
 data AudioRate
